Fixed problem where environment would not be provisioned on first run.
According to https://github.com/mitchellh/vagrant/issues/2421, the default behavior was changed since Vagrant 1.3.0 on is to not provision new boxes after the first run unless you specify either "vagrant up --provision" (to bring it up after a "vagrant destroy") or "vagrant reload --provision" or "vagrant provision" for a running instance.

In order to fix it, I removed the 'if $rvm_installed == "true"' condition. The purpose of that condition to begin with, is to avoid an error that comes up when running puppet with --noop (as specified at https://github.com/blt04/puppet-rvm/tree/master under the Troubleshooting / FAQ section). Since $rvm_installed is evaluated at the beginning of each puppet run, that fix however has the side effect that if you use this in your manifest, you will need to run puppet twice to fully configure RVM.